refactor(bitbucketCloud): extracts function getAuthorizationHeader into helper module
